+++ CHANGELOG.md
@@ -10,6 +10,10 @@
     * `translated` -> `translated` with no sub-state.
   * Let `xliffSync.fileType` = `xlf2` work with file-extension `xlf`.
 
+### Thank You
+
+* **[antpyykk](https://github.com/antpyykk)** for filing the issue with states for XLIFF 2.0 files (GitHub issue [#50](https://github.com/rvanbekkum/vsc-xliff-sync/issues/50))
+
 ## [0.4.0] 02-08-2020
 
 * New setting `xliffSync.syncCrossWorkspaceFolders` which can be used to set that the extension should synchronize from one single base file (`xliffSync.baseFile`) to the translation files in all workspace folders (**Default**: `false`) (GitHub issue [#48](https://github.com/rvanbekkum/vsc-xliff-sync/issues/48)).
@@ -27,6 +31,10 @@
 * XLIFF Sync snippets for the "Parse from Developer Note" feature.
   * You can configure for which programming languages the snippets should be available with setting `xliffSync.enableSnippetsForLanguages`. Currently only the "AL Language" is supported with snippets: `tcaptionwithtranslation`, `tcommentwithtranslation`, `toptioncaptionwithtranslation`, `tpromotedactioncategorieswithtranslation`, `tlabelwithtranslation` and `ttooltipwithtranslation` snippets.
   * You can configure a default target language that should be used by the snippets with setting `xliffSync.snippetTargetLanguage`.
+
+### Thank You
+
+* **[GregoryAA](https://github.com/GregoryAA)** for requesting the _Parse from Developer Note_ feature enhancements. (GitHub issue [#43](https://github.com/rvanbekkum/vsc-xliff-sync/issues/43))
 
 ## [0.3.7] 18-05-2020